Sun Rework
Instead of buying cards and having the chance of the user take double damage next turn.

now it buffs the user's damage by x1.3 in its next turn, only if the opponent by the end of its turn when this card is played have its energy depleted.

now, The Star deals 20 damage instead of 23 to account The Sun's rework since they can combo. and Judgement deals 2 instead of 1 for the same reason(other wise would be pointless).

and the expensive pack now costs 12 instead of 15.